moving the Either stuff into its own module
