+++ hgeometry/hgeometry.cabal
@@ -247,6 +247,7 @@ library point
        HGeometry.Point.PointF
        HGeometry.Point.EuclideanDistance
        HGeometry.Point.Orientation
+       HGeometry.Point.Either
 
     other-modules:
         HGeometry.Point.Type
